fix: 🐛 Fix Anime themes missing song crashing media page
